AMB-447 separate happy path spine tests from proxy tests, and add auth test for testing ASID
