--- api_tests/tests/reasonable_adjustment_flag/request_bodies/POST_Consent.json
@@ -1,50 +1,0 @@
-{
-  "resourceType": "Consent",
-  "meta": {
-    "profile": [
-      "https://fhir.nhs.uk/STU3/StructureDefinition/RARecord-Consent-1"
-    ]
-  },
-  "extension": [
-    {
-      "url": "  ",
-      "valueCodeableConcept": {
-        "coding": [
-          {
-            "system": "https://fhir.nhs.uk/STU3/CodeSystem/RARecord-ProxyRole-1",
-            "code": "003",
-            "display": "Lasting power of attorney for personal welfare or court-appointed deputy or equivalent (Mental Capacity Act 2005)"
-          }
-        ]
-      }
-    }
-  ],
-  "status":"active",
-  "category": [
-    {
-      "coding": [
-        {
-          "system": "https://fhir.nhs.uk/STU3/CodeSystem/RARecord-FlagCategory-1",
-          "code": "NRAF",
-          "display": "Reasonable Adjustments Flag"
-        }
-      ]
-    }
-  ],
-  "patient": {
-    "reference": "demographics.spineservices.nhs.uk/STU3/Patient/5900008142"
-  },
-  "policy": [
-    {
-      "authority": "https://www.gov.uk/",
-      "uri": "https://www.gov.uk/government/uploads/system/uploads/attachment_data/file/535024/data-security-review.pdf"
-    }
-  ],
-  "purpose": [
-    {
-      "system": "https://fhir.nhs.uk/STU3/CodeSystem/RARecord-ConsentPurpose-1",
-      "code": "RACONSENT",
-      "display": "Reasonable Adjustments - Consent to record Reasonable Adjustments"
-    }
-  ]
-}
