--- api_tests/tests/reasonable_adjustment_flag/request_bodies/POST_List.json
@@ -1,1 +1,0 @@
-{}
