--- api_tests/tests/reasonable_adjustment_flag/request_bodies/POST_removerarecord.json
@@ -1,30 +1,0 @@
-{
-  "resourceType": "Parameters",
-  "parameter": [
-    {
-      "name": "removerarecord",
-      "part": [
-        {
-          "name": "nhsNumber",
-          "valueString": "5900008142"
-        },
-        {
-          "name": "removalReason",
-          "valueCodeableConcept": {
-            "coding": [
-              {
-                "system": "https://fhir.nhs.uk/STU3/CodeSystem/CodeSystem-RARecord-RemovalReason-1",
-                "code": "DoesntApply",
-                "display": "The Reasonable Adjustment Flag no longer applies to the patient"
-              }
-            ]
-          }
-        },
-        {
-          "name": "supportingComment",
-          "valueString": "No longer applies"
-        }
-      ]
-    }
-  ]
-}
