--- api_tests/tests/reasonable_adjustment_flag/request_bodies/PUT_Flag.json
@@ -1,122 +1,0 @@
-{
-  "resourceType": "Flag",
-  "id": "2acb0536-0a8f-48c9-8a2f-6ee82860f186",
-  "fhir_comments": [
-    " xml example fileUpdateExample-UpdateFlagRequest.xml "
-  ],
-  "meta": {
-    "versionId": "aa755bd6-2be9-4971-972a-6724879c5cb1",
-    "lastUpdated": "2018-07-24T10:01:00+00:00",
-    "profile": [
-      "https://fhir.nhs.uk/STU3/StructureDefinition/RARecord-Flag-1"
-    ]
-  },
-  "contained": [
-    {
-      "resourceType": "Provenance",
-      "id": "43124f67-b09d-453d-b889-a5e2e8780b9e",
-      "meta": {
-        "profile": [
-          "https://fhir.nhs.uk/STU3/StructureDefinition/RARecord-Provenance-1"
-        ]
-      },
-      "target": [
-        {
-          "reference": "Flag/2acb0536-0a8f-48c9-8a2f-6ee82860f186"
-        }
-      ],
-      "recorded": "2018-07-24T10:05:33+00:00",
-      "agent": [
-        {
-          "role": [
-            {
-              "coding": [
-                {
-                  "system": "https://fhir.nhs.uk/STU3/CodeSystem/CareConnect-SDSJobRoleName-1",
-                  "code": "R1974",
-                  "display": "Community Learning Disabilities Nurse"
-                }
-              ]
-            }
-          ],
-          "whoReference": {
-            "reference": "https://sds.spineservices.nhs.uk/STU3/Practitioner/4tr6ee6a9",
-            "display": "Nurse N"
-          },
-          "onBehalfOfReference": {
-            "reference": "https://directory.spineservices.nhs.uk/STU3/Organization/a3e5i7",
-            "display": "Some Hospital Learning Disability Support Department"
-          }
-        }
-      ]
-    }
-  ],
-  "extension": [
-    {
-      "url": "https://fhir.nhs.uk/STU3/StructureDefinition/Extension-RARecord-Provenance-1",
-      "extension": [
-        {
-          "url": "created",
-          "valueReference": {
-            "reference": "#43124f67-b09d-453d-b889-a5e2e8780b9e"
-          }
-        }
-      ]
-    },
-    {
-      "url": "https://fhir.nhs.uk/STU3/StructureDefinition/Extension-RARecord-RemovalReason-1",
-      "extension": [
-        {
-          "url": "removalReason",
-          "valueCodeableConcept": {
-            "coding": [
-              {
-                "system": "https://fhir.nhs.uk/STU3/CodeSystem/RARecord-RemovalReason-1",
-                "code": "Error",
-                "display": "The Reasonable Adjustment Flag was created in error"
-              }
-            ]
-          }
-        },
-        {
-          "url": "supportingComment",
-          "valueString": "Requires Large Print rather than Easy Read"
-        }
-      ]
-    },
-    {
-      "url": "https://fhir.nhs.uk/STU3/StructureDefinition/Extension-RARecord-AdjustmentCategory-1",
-      "valueCodeableConcept": {
-        "coding": [
-          {
-            "system": "https://fhir.nhs.uk/STU3/CodeSystem/CodeSystem-RARecord-AdjustmentCategories-1",
-            "code": "comms",
-            "display": "Communication"
-          }
-        ]
-      }
-    }
-  ],
-  "status": "inactive",
-  "category": {
-    "coding": [
-      {
-        "system": "https://fhir.nhs.uk/STU3/CodeSystem/CodeSystem-RARecord-FlagCategory-1",
-        "code": "reasonable adjustments flag",
-        "display": "Reasonable Adjustments Flag"
-      }
-    ]
-  },
-  "code": {
-    "coding": [
-      {
-        "system": "https://fhir.nhs.uk/STU3/CodeSystem/CodeSystem-RARecord-AccessibleInformationAdjustments-1",
-        "code": "requiresinformationineasyread",
-        "display": "Requires information in Easyread"
-      }
-    ]
-  },
-  "subject": {
-    "reference": "demographics.spineservices.nhs.uk/STU3/Patient/5900008142"
-  }
-}
